requirements.txt
langchain-community==0.3.7
langchain-core==0.3.20
langchain-google-genai==2.0.5